feat: migrate to api route handler
